Fixed Ctrl key combinations by reassigning the scan codes
Ctrl key combinations were not being moved with the letter keys. For
example, Ctrl+C was still in the QWERTY location. Fixed this in the
Keyboard Layout Creator by reassigning the scan code to virtual key
mapping in the advanced view. Only changed the US versions, as I don't
have a UK keyboard to test with.