Add space after Workman dead true comma
Normally, when the user hits the dead key and then the space bar, it
produces a comma. Here it produce a comma plus a space. This makes
normal typing behave as expected, while getting all the benefits of the
dead key. On those rare occasion where it's necessary to type a comma
without a space following it, the backspace will do.